chore: move bus builder to different files
