+++ processor/src/chiplets/aux_trace/bus/messages.rs
